Add metadata to contact layout page
