+++ app/contact/page.tsx
@@ -1,20 +1,8 @@
 "use client";
 
 import ContactForm from "@/components/ContactForm";
-import { Metadata } from "next";
 import { useSearchParams } from "next/navigation";
 import { useEffect, useState } from "react";
-
-export const metadata: Metadata = {
-	title: "Contact Us",
-	description:
-		"Get in touch with the Trip Report team. We're here to help with any questions or feedback you may have.",
-	openGraph: {
-		title: "Contact Trip Report",
-		description: "Reach out to us with your questions or feedback.",
-		type: "website",
-	},
-};
 
 export default function ContactPage() {
 	const searchParams = useSearchParams();
